pyproject.toml
[project]
name = "assinment-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.4",
    "streamlit>=1.42.2",
]
